fix: create custom hook, moved logic to custom hook
